Adjusting reusability of PetTaskModel
