+++ petJournal/petJournal/Commons/UIComponents/PetTaskCard.swift
@@ -1,19 +1,8 @@
 import SwiftUI
 
-// MARK: - Models
-struct PetTask: Identifiable {
-    let id = UUID()
-    let title: String
-    let schedule: String
-    let description: String
-    let petImages: [Image]
-    let accentColor: Color
-    let backgroundIcon: Image
-}
-
 // MARK: - Protocols
 protocol PetTaskCardDisplayLogic {
-    var task: PetTask { get }
+    var task: PetTaskModel { get }
     var isExpanded: Bool { get set }
     func toggleExpansion()
 }
@@ -21,9 +10,9 @@ protocol PetTaskCardDisplayLogic {
 // MARK: - Presenter
 class PetTaskCardPresenter: ObservableObject, PetTaskCardDisplayLogic {
     @Published var isExpanded: Bool = false
-    let task: PetTask
+    let task: PetTaskModel
     
-    init(task: PetTask) {
+    init(task: PetTaskModel) {
         self.task = task
     }
     
@@ -158,28 +147,9 @@ struct PetTaskCard: View {
 // MARK: - Preview
 struct PetTaskCard_Previews: PreviewProvider {
     static var previews: some View {
-        let tasks = [
-            PetTask(
-                title: "Carprofeno",
-                schedule: "Manhã e noite",
-                description: "Anti-inflamatório não esteroide para alívio da dor e inflamação.\n\nAqui tem mais informação para ser lida camarada! Você pode ser até uma informação bem detalhada com todo cuidado que seu Pet merece <3",
-                petImages: PetModel.mockPetImages,
-                accentColor: Color.theme.petSecondary500,
-                backgroundIcon: Image(asset: .medicine)
-            ),
-            PetTask(
-                title: "Consulta médica",
-                schedule: "15/06 às 14:00",
-                description: "Check-up anual\n\nAqui tem mais informação para ser lida camarada! Você pode ser até uma informação bem detalhada com todo cuidado que seu Pet merece <3",
-                petImages: PetModel.mockPetImages,
-                accentColor: Color.theme.petCerise,
-                backgroundIcon: Image(asset: .vetAppointment)
-            )
-        ]
-        
         return ScrollView {
             VStack(spacing: 20) {
-                ForEach(tasks) { task in
+                ForEach(PetTaskModel.sampleTasks) { task in
                     PetTaskCard(presenter: PetTaskCardPresenter(task: task))
                 }
             }
